cleaned up code, added various functionality, including file permissions (one can't delete another token's file)

+++ scripts/delete.php
@@ -9,13 +9,20 @@ if (! isset($token) || ! check_valid_token($token))
 $filename = $_GET['fn'];
 $CONFIG = include "config.php";
 
+if (! is_file_owner($token, $filename))
+    die(json_encode(array(
+        'error' => 'no valid filename provided',
+        'debug' => ''
+    )));
+
 $target = $CONFIG['filestore_dir'] . "/" . $filename;
 if (! unlink($target))
     die(json_encode(array(
         'error' => 'no valid filename provided',
         'debug' => 'file either already deleted or never existed'
     )));
-$result = delete_file($token, $filename);
+
+$result = delete_file($filename);
 die(json_encode(array(
     'success' => '1',
     'debug' => $result
